performance tweaks

+++ src/main/java/info/rmarcus/horton/RedirectList.java
@@ -2,14 +2,14 @@ package info.rmarcus.horton;
 
 public class RedirectList<K, V> extends KVPair<K, V> {
 
-	public short[] redirectList;
+	public byte[] redirectList;
 	public static final int REDIRECT_TABLE_SIZE = 8;
 
 	
 	public RedirectList() {
 		super(null, null);
 		
-		redirectList = new short[REDIRECT_TABLE_SIZE];
+		redirectList = new byte[REDIRECT_TABLE_SIZE];
 		for (int i = 0; i < REDIRECT_TABLE_SIZE; i++) {
 			redirectList[i] = -1;
 		}
